requirements.txt
tensorflow
transformers
gradio
python-dotenv